Refactor user repository imports
